post auth added

+++ app/Http/Controllers/Api/v1/PostController.php
@@ -19,28 +19,82 @@ class PostController extends Controller
     	return new PostResourceCollection(Post::paginate());
     }
 
-    public function store(Request $request)
+    public function update(Request $request, Post $post)
     {
-    	$request->validate([
-    		'userId' => 'required',
-    		'productId' => 'required',
-    		'postTitle' => 'required',
-    		'postDescription' =>'required'
-    	]);
+        
+        try {
+            $user = auth()->userOrFail();
+        } catch (\Tymon\JWTAuth\Exceptions\UserNotDefinedException $e) {
+            return response()->json(['error'=> $e->getMessage()]);
+        }
 
-    	$post = post::create($request->all());
-    	return new PostResource($post);
-    }
+        if($post->userId != $user->id){
+            return response()->json(['error'=> "Unable to modify other user's post"]);
+        }
 
-    public function update(Request $request, Post $post):PostResource
-    {
+        if($request->userId != $user->id){
+            return response()->json(['error'=> "Unable to modify post for other user"]);
+        }
+
     	$post->update($request->all());
     	return new PostResource($post);
     }
 
     public function destroy(Post $post)
     {
+        try {
+            $user = auth()->userOrFail();
+        } catch (\Tymon\JWTAuth\Exceptions\UserNotDefinedException $e) {
+            return response()->json(['error'=> $e->getMessage()]);
+        }
+
+        if($post->userId != $user->id){
+            return response()->json(['error'=> "Unable to delete other user's post"]);
+        }
     	$post->delete();
     	return response()->json([]);
+    } 
+
+    public function self()
+    {
+
+       try {
+            $user = auth()->userOrFail();
+        } catch (\Tymon\JWTAuth\Exceptions\UserNotDefinedException $e) {
+            return response()->json(['error'=> $e->getMessage()]);
+        } 
+        $post = Post::where('userId',$user->id)->first();
+        return $post;
+    } 
+
+    public function store(Request $request)
+    {
+        $request->validate([
+            'userId' => 'required',
+            'productId' => 'required',
+            'postTitle' => 'required',
+            'postDescription' =>'required'
+        ]);
+
+        try {
+            $user = auth()->userOrFail();
+        } catch (\Tymon\JWTAuth\Exceptions\UserNotDefinedException $e) {
+            return response()->json(['error'=> $e->getMessage()]);
+        } 
+        if($request->userId != $user->id){
+            return response()->json(['error'=> "Unable to add post for other user"]);
+        }
+        $post = post::create($request->all());
+        return new PostResource($post);
     }
+    
 }
+
+
+
+
+
+
+
+
+
